refactor: Tier 1 command consolidation — trash subgroup, maintain nesting, remove gist-list
- Remove broken `gist-list` command (66 LOC)
- Consolidate `trash`, `restore`, `purge` into `emdx trash` subgroup
- Nest `cleanup`, `cleanup-dirs`, `analyze` under `emdx maintain`
- Update tests and docs to match new command paths
- 28 top-level commands (down from 35)
