+++ emdx/main.py
@@ -100,13 +100,13 @@ register_lazy_commands(get_lazy_subcommands(), get_lazy_help())
 from emdx.commands.core import app as core_app
 from emdx.commands.browse import app as browse_app
 from emdx.commands.tags import app as tag_app
+from emdx.commands.trash import app as trash_app
 from emdx.commands.executions import app as executions_app
 from emdx.commands.tasks import app as tasks_app
 from emdx.commands.groups import app as groups_app
 from emdx.commands.prime import prime as prime_command
 from emdx.commands.status import status as status_command
 from emdx.ui.gui import gui as gui_command
-from emdx.commands.analyze import app as analyze_app
 from emdx.commands.maintain import app as maintain_app
 from emdx.commands.gist import app as gist_app
 
@@ -145,6 +145,9 @@ for command in gist_app.registered_commands:
 # Tag commands (subcommand group: emdx tag <subcommand>)
 app.add_typer(tag_app, name="tag", help="Manage document tags")
 
+# Trash commands (subcommand group: emdx trash <subcommand>)
+app.add_typer(trash_app, name="trash", help="Manage deleted documents")
+
 # Add executions as a subcommand group
 app.add_typer(executions_app, name="exec", help="Manage Claude executions")
 
@@ -154,13 +157,8 @@ app.add_typer(tasks_app, name="task", help="Task management")
 # Add groups as a subcommand group
 app.add_typer(groups_app, name="group", help="Organize documents into hierarchical groups")
 
-# Add analyze commands (maintain_app has multiple commands like cleanup, cleanup-dirs)
-for command in analyze_app.registered_commands:
-    app.registered_commands.append(command)
-
-# Add maintain commands (maintain_app has maintain, cleanup, cleanup-dirs)
-for command in maintain_app.registered_commands:
-    app.registered_commands.append(command)
+# Add maintain as a subcommand group (includes maintain, cleanup, cleanup-dirs, analyze)
+app.add_typer(maintain_app, name="maintain", help="Maintenance and analysis tools")
 
 # Add the prime command for Claude session priming
 app.command(name="prime")(prime_command)
